Move downloading out of main.lua
downloadBook was 244 lines, the largest thing left after discovery went, and it
came with two functions only it uses: _fetchDetailsThenDownload, and the
_usableFormat guard that stopped a filename being built out of "N/A". 308 lines
together, now zlibrary/download.lua.

Moved verbatim, the same way as discovery: the module functions take a
parameter literally named `self`, so the bodies are byte-for-byte what they
were -- 296 lines, checked against git. The plugin methods they call, login and
resetDownloadQuotaCache and each other, resolve on the instance exactly as
before. main.lua keeps two one-line methods delegating here, because
bookdetails_dialog calls downloadBook on the plugin instance and the retry and
detail-fetch paths recurse through it.

main.lua: 1794 -> 1158 lines across the two extractions. No translatable string
moved, gained or lost, and the module graph stays acyclic.

The extraction harness now covers both modules from a table rather than naming
one, so a third costs a few lines. Writing that found a fault in its own
unbound-name scan: download.lua's comments explain why the child owns the
socket and why an error cannot be read out of socket.http, and the scanner read
those as calls and reported `socket` as a missing dependency. It strips
comments now.

What remains in main.lua is coupled and should stay: addToMainMenu is the menu
definition, showMyBooksDialog calls eight sibling methods, displaySearchResults
owns six fields of search state. Splitting those means inventing somewhere for
that state to live, which is a change rather than a move, and worth far more
care than either of these.
